refactor: move server to repo root, rename module to gitstats, delete client
