+++ .github/workflows/ci.yaml
@@ -11,9 +11,8 @@ jobs:
 
       - uses: actions/setup-go@v5
         with:
-          go-version-file: server/go.mod
-          cache-dependency-path: server/go.sum
+          go-version-file: go.mod
+          cache-dependency-path: go.sum
 
       - name: test
-        working-directory: server
         run: go test ./...
